requirements.txt
EasyProcess==0.2.5
numpy==1.16.3
Pillow==6.0.0
PyAutoGUI==0.9.42
PyGetWindow==0.0.5
PyMsgBox==1.0.6
PyRect==0.1.4
pyscreenshot==0.5.1
PyScreeze==0.1.21
python3-xlib==0.15
PyTweening==1.0.3
